+++ src/Message/Response/Stringable.php
@@ -11,7 +11,14 @@ use Innmind\Http\{
     Headers,
     Header,
 };
-use Innmind\Filesystem\File\Content;
+use Innmind\Filesystem\{
+    File\Content,
+    Chunk,
+};
+use Innmind\Immutable\{
+    Sequence,
+    Str,
+};
 
 /**
  * @psalm-immutable
@@ -23,6 +30,14 @@ final class Stringable implements ResponseInterface
     public function __construct(ResponseInterface $response)
     {
         $this->response = $response;
+    }
+
+    /**
+     * @psalm-pure
+     */
+    public static function of(ResponseInterface $response): self
+    {
+        return new self($response);
     }
 
     public function statusCode(): StatusCode
@@ -45,27 +60,32 @@ final class Stringable implements ResponseInterface
         return $this->response->body();
     }
 
+    public function asContent(): Content
+    {
+        $status = Str::of("HTTP/%s %s %s\n")->sprintf(
+            $this->protocolVersion()->toString(),
+            $this->statusCode()->toString(),
+            $this->statusCode()->reasonPhrase(),
+        );
+        $headers = $this
+            ->headers()
+            ->all()
+            ->sort(static fn($a, $b) => $b->name() <=> $a->name())
+            ->map(static fn($header) => $header->toString())
+            ->map(Str::of(...))
+            ->map(static fn($header) => $header->append("\n"));
+        $body = (new Chunk)($this->body());
+
+        return Content\Chunks::of(
+            Sequence::lazyStartingWith($status)
+                ->append($headers)
+                ->add(Str::of("\n"))
+                ->append($body),
+        );
+    }
+
     public function toString(): string
     {
-        $headers = $this->headers()->reduce(
-            [],
-            static function(array $headers, Header $header): array {
-                $headers[] = $header;
-
-                return $headers;
-            },
-        );
-        $headers = \array_map(
-            static fn(Header $header): string => $header->toString(),
-            $headers,
-        );
-        $headers = \implode("\n", $headers);
-
-        return <<<RAW
-HTTP/{$this->protocolVersion()->toString()} {$this->statusCode()->toString()} {$this->statusCode()->reasonPhrase()}
-$headers
-
-{$this->body()->toString()}
-RAW;
+        return $this->asContent()->toString();
     }
 }